Merge upstream/dev into doris branch. Upstream brings scan scheduler
improvements, pipeline error recovery, nvtx3 profiling, and new Claude
Code skills.

Doris-specific fixes on top of merge:
- Call QueryBegin explicitly in sirius_engine::execute() — DuckDB's
  ClientContextState callback doesn't fire in the Doris embedded path,
  leaving task_creator's _client_context null (SIGSEGV on PARQUET_SCAN)
- Pin cuda-version to 13.1.* — cudarc 0.19.4 with CUDA 13.2 demands
  cuCoredumpDeregisterCompleteCallback which the driver doesn't expose
- Bump cudarc 0.19 → 0.19.4 for CUDA 13 support
- Fix duplicate [environments] section in pixi.toml
- Route UNION ALL exchange through GPU SQL path when GPU tables exist
- Add substrait include paths to CMakeLists.txt for from_substrait

+++ src/expression_executor/gpu_expression_translator.cpp
@@ -70,7 +70,7 @@ std::optional<expr_ref> gpu_expression_translator::add_join_condition(
         cudf::ast::ast_operator::GREATER_EQUAL, *left_expr, *right_expr);
     default:
       SIRIUS_LOG_DEBUG("[expression_translator] Unsupported join condition comparison type: {}",
-                       condition.comparison);
+                       static_cast<int>(condition.comparison));
       return std::nullopt;
   }
 }
@@ -194,7 +194,7 @@ std::optional<expr_ref> gpu_expression_translator::add_expression(
                                                      *child_expr);
     default: {
       SIRIUS_LOG_DEBUG("[expression_translator] Unsupported cast type_id: {}",
-                       cudf_return_type.id());
+                       static_cast<int>(cudf_return_type.id()));
       return std::nullopt;
     }
   }
@@ -327,7 +327,8 @@ std::optional<expr_ref> gpu_expression_translator::add_expression(
         _resource_ref);
     }
     default: {
-      SIRIUS_LOG_DEBUG("[expression_translator] Unsupported constant type_id: {}", cudf_type.id());
+      SIRIUS_LOG_DEBUG("[expression_translator] Unsupported constant type_id: {}",
+                       static_cast<int>(cudf_type.id()));
       return std::nullopt;
     }
   }
